+++ module/news/ressource/defaultdata.php
@@ -5,13 +5,16 @@ class init extends search {
         'feedsLabel'         => '',
         'itemsperPage'       => 8,
         'itemsperCol'        => 12,
-        'height'             => 1000,
+        'height'             => -1,
         'versionData'        => '3.4'
     ];
-    /*
+
     public static $defaultTheme = [
-        'itemsHeight'        => 'auto',
-        'itemsBlur'         => '0%'
+        'style'               => '',
+        'borderStyle'         => 'none',
+        'borderColor'         => 'rgba(33, 34, 35, 1)',
+        'backgroundColor'     => 'rgba(255, 255, 255, 1)',
+        'borderWidth'          => '0'
     ];
-    */
+
 }