+++ module/news/view/config/config.php
@@ -40,9 +40,8 @@
 				<div class="row">
 					<div class="col4">
 						<?php echo template::select('newsConfigItemsperCol', $module::$columns, [
-							'label' => 'Pagination',
-							'selected' => $this->getData(['module', $this->getUrl(0),'config', 'itemsperCol']),
-							'help' => 'Nombre de colonnes par page'
+							'label' => 'Nombre de colonnes',
+							'selected' => $this->getData(['module', $this->getUrl(0),'config', 'itemsperCol'])
 						]); ?>
 					</div>
 					<div class="col4">
@@ -53,9 +52,45 @@
 					</div>
 					<div class="col4">
 						<?php echo template::select('newsConfigHeight', $module::$height, [
-							'label' => 'Hauteur',
-							'selected' => $this->getData(['module', $this->getUrl(0),'config', 'height']),
-							'help' => 'Limite la hauteur de l\'article, cette option est utile lorsque la pagination en colonnes est activée.'
+							'label' => 'Abrégé de l\'article',
+							'selected' => $this->getData(['module', $this->getUrl(0),'config', 'height'])
+						]); ?>
+					</div>
+				</div>
+			</div>
+		</div>
+	</div>
+	<div class="row">
+		<div class="col12">
+			<div class="block">
+				<h4>Thème du module</h4>
+				<div class="row">
+					<div class="col3">
+						<?php echo template::select('newsThemeBorderStyle', $module::$borderStyle, [
+							'label' => 'Bordure',
+							'selected' => $this->getData(['module', $this->getUrl(0),'theme', 'borderStyle'])
+						]); ?>
+					</div>
+					<div class="col3">
+						<?php echo template::select('newsThemeBorderWidth', $module::$borderWidth, [
+							'label' => 'Epaisseur',
+							'selected' => $this->getData(['module', $this->getUrl(0),'theme', 'borderWidth'])
+						]); ?>
+					</div>
+					<div class="col3">
+						<?php echo template::text('newsThemeBorderColor', [
+							'class' => 'colorPicker',
+							'help' => 'Couleur visible en l\'absence d\'une image.<br />Le curseur horizontal règle le niveau de transparence.',
+							'label' => 'Couleur de la bordure',
+							'value' => $this->getData(['module', $this->getUrl(0),'theme', 'borderColor'])
+						]); ?>
+					</div>
+					<div class="col3">
+						<?php echo template::text('newsThemeBackgroundColor', [
+							'class' => 'colorPicker',
+							'help' => 'Couleur visible en l\'absence d\'une image.<br />Le curseur horizontal règle le niveau de transparence.',
+							'label' => 'Couleur du fond',
+							'value' => $this->getData(['module', $this->getUrl(0),'theme', 'backgroundColor'])
 						]); ?>
 					</div>
 				</div>